Adding API tests for SWAPI
First commit for some tests that aim to serve as a first
approach for using cy.request command.
We are sending some requests to the SWAPI api that you can
find here: https://swapi.dev/documentation
We are sending requests to the planets and people endpoints and
we are doing some assertions on the responses, extracting values
of the responses and doing another request using the extacted value.